refactor(storyreader): create story selector for both text and live2d reader
